+++ internal/app/wwctl/node/console/power.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/hpcng/warewulf/internal/pkg/hostlist"
 	"github.com/hpcng/warewulf/internal/pkg/node"
 	"github.com/hpcng/warewulf/internal/pkg/power"
 	"github.com/hpcng/warewulf/internal/pkg/wwlog"
-	"github.com/hpcng/warewulf/pkg/hostlist"
 	"github.com/spf13/cobra"
 )